+++ casadocodigo/src/main/java/br/com/casadocodigo/loja/daos/ProdutoDAO.java
@@ -26,7 +26,7 @@ public class ProdutoDAO {
 		
 	}
 
-	public Produto find(long id) {
+	public Produto find(int id) {
 		return manager.createQuery("select distinct(p) from Produto p join fetch p.precos precos where p.id = :id", Produto.class)
 				.setParameter("id", id)
 				.getSingleResult();
